+++ AzureDevOpsBackup/Program.cs
@@ -107,7 +107,7 @@ namespace AzureDevOpsBackup
             bool noProjectsToBackup = false;
             bool isOutputFolderContainFiles = false;
 
-            // Get key to use
+            // Get key to use for encryption
             var key = SecureArgumentHandlerToken.GetComputerId();
 
             // Get application data to later use in tool
@@ -152,7 +152,7 @@ namespace AzureDevOpsBackup
             string[] requiredArgs = { "--token", "--org", "--backup", "--server", "--port", "--from", "--to" };
             
             // Check if parameters have been provided and Contains one of
-            if (args.Length == 0 || args.Contains("--help") || args.Contains("/h") || args.Contains("/?") || args.Contains("/info") || args.Contains("/about") || args.Contains("--tokentofile") || args.Contains("/tokentofile"))
+            if (args.Length == 0 || args.Contains("--help") || args.Contains("/h") || args.Contains("/?") || args.Contains("/info") || args.Contains("/about") || args.Contains("--tokenfile"))
             {
                 // If none arguments
                 if (args.Length == 0)
@@ -201,23 +201,13 @@ namespace AzureDevOpsBackup
                 }
                 
                 // If wants to save token data as file for encryption
-                if (args.Contains("--tokenfile") || args.Contains("/tokenfile"))
+                if (args.Contains("--tokenfile"))
                 {
                     // Get data fon console
                     string tokentoencrypt = args[Array.IndexOf(args, "--tokenfile") + 1];
                     
                     // Encrypt data
                     SecureArgumentHandlerToken.EncryptAndSaveToFile(key, tokentoencrypt);
-
-                    // Decrypt and show data in .bin file
-                    //byte[] data = File.ReadAllBytes(Files.TokenFilePath);
-                    //string text = Encoding.UTF8.GetString(data);
-                    //Console.WriteLine($"Encrypted string = {text}");
-                    //Console.WriteLine($"Encrypted string = {SecureArgumentHandlerToken.DecryptFromFile(key)}");
-
-                    // Read the token information from the -tokentofile
-                    //var decryptedString = SecureArgumentHandlerToken.DecryptFromFile(key);
-                    //Console.WriteLine($"Decrypted string = {decryptedString}");
 
                     Message($"Saved information about token to file - Exciting {Globals.AppName}, v." + Globals._vData + " by " + Globals._companyName + "!", EventType.Information, 1000);
 
@@ -291,8 +281,10 @@ namespace AzureDevOpsBackup
                         {
                             // Read the token information from the -tokentofile
                             token = SecureArgumentHandlerToken.DecryptFromFile(key);
-                            Console.WriteLine($"Decrypted string for token = {token}");
-                            Console.ReadKey();
+#if DEBUG
+                            //Console.WriteLine($"Decrypted string for token = {token}");
+                            //Console.ReadKey();
+#endif
                         }
 
                         // Encrypt the value
@@ -848,13 +840,6 @@ namespace AzureDevOpsBackup
                                             noProjectsToBackup = true;
                                         }
                                     }
-                                    
-                                    // Log
-                                    //Message("Getting information about Git repository: " + repo.Name, EventType.Information, 1000);
-                                    //Console.WriteLine("Getting information about Git repository: " + repo.Name);
-                                    //var clientItems = new RestClient(baseUrl + "_apis/git/repositories/" + repo.Id + "/items?recursionlevel=full&" + version + "&versionDescriptor.versionType=Branch&versionDescriptor.version=" + branchName);
-                                    //var clientItems = new RestClient(baseUrl + "_apis/git/repositories/" + repo.Id + "/items?recursionlevel=full&" + version);
-                                    //var requestItems = new RestRequest(Method.GET);
                                 }
                             }
 
@@ -987,78 +972,7 @@ namespace AzureDevOpsBackup
                             }
 
                             // Cleanup old log files
-                            string[] oldfiles = Directory.GetFiles(Files.LogFilePath);
-
-                            // Log
-                            Message("Checking for old log file(s) to cleanup...", EventType.Information, 1000);
-                            Console.ForegroundColor = ConsoleColor.Yellow;
-                            Console.WriteLine("\nChecking for old log file(s) to cleanup...");
-                            Console.ResetColor();
-
-                            // Loop all files in folder
-                            foreach (string file in oldfiles)
-                            {
-                                FileInfo fi = new FileInfo(file);
-
-                                // Get all last access time back in time
-                                if (fi.LastAccessTime < DateTime.Now.AddDays(-30))
-                                {
-                                    try
-                                    {
-                                        // Do work
-                                        fi.Delete();
-
-                                        // Log
-                                        Message("Deleted old log file: " + fi, EventType.Information, 1000);
-                                        Console.ForegroundColor = ConsoleColor.Yellow;
-                                        Console.WriteLine("Deleted old log file: " + fi);
-                                        Console.ResetColor();
-
-                                        // Set status
-                                        oldLogfilesToDelete = true;
-                                        oldLogFilesToDeleteCount++;
-                                    }
-                                    catch (UnauthorizedAccessException)
-                                    {
-                                        Message("Unable to delete old log file: " + fi + ". Make sure the account you use to run this tool has delete rights to this location.", EventType.Error, 1001);
-                                        Console.ForegroundColor = ConsoleColor.Red;
-                                        Console.WriteLine("Unable to delete old log file: " + fi + ". Make sure the account you use to run this tool has delete rights to this location.");
-                                        Console.ResetColor();
-
-                                        // Count errors
-                                        Globals._errors++;
-                                    }
-                                    catch (Exception ex)
-                                    {
-                                        // Log
-                                        Message("Sorry, we are unable to delete old log file: " + fi + "Error: " + ex, EventType.Error, 1001);
-                                        Console.ForegroundColor = ConsoleColor.Red;
-                                        Console.WriteLine("Sorry, we are unable to delete old log file: " + fi + "Error: " + ex);
-                                        Console.ResetColor();
-
-                                        // Count errors
-                                        Globals._errors++;
-                                    }
-                                }
-                            }
-
-                            // Check if there is old log files to delete
-                            if (oldLogfilesToDelete)
-                            {
-                                // Log
-                                Message($"There was {oldLogFilesToDeleteCount} old log files to delete (-30 days)", EventType.Information, 1000);
-                                Console.ForegroundColor = ConsoleColor.Green;
-                                Console.WriteLine($"There was {oldLogFilesToDeleteCount} old log files to delete (-30 days)");
-                                Console.ResetColor();
-                            }
-                            else
-                            {
-                                // Log
-                                Message("No old log files to delete (-30 days)", EventType.Information, 1000);
-                                Console.ForegroundColor = ConsoleColor.Green;
-                                Console.WriteLine("No old log files to delete (-30 days)");
-                                Console.ResetColor();
-                            }
+                            CleanupLog.CleanupLogs();
 
                             // Get email status text from job status
                             if (isBackupOk)
@@ -1538,6 +1452,10 @@ namespace AzureDevOpsBackup
                         Console.WriteLine(isOutputFolderContainFilesStatusText);
                         Console.ResetColor();
                         Message(isOutputFolderContainFilesStatusText, EventType.Information, 1000);
+
+
+                        // Count backups in backup folder
+                        Backups.CountCurrentNumersOfBackup(outBackupDir);
 
                         // Log
                         Message($"Getting status for tasks for email report is done", EventType.Information, 1000);