+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [1.0.5.5] - xx-x-2023
 
 ### Add
-- 
+- Added function to count current numers of backups in backup folder and report to email report
 
 ### Changed
 - Some changes to help text